requirements.txt
absl-py==0.6.1
appnope==0.1.0
asn1crypto==0.24.0
astor==0.7.1
attrs==18.2.0
Automat==0.7.0
backcall==0.1.0
bleach==1.5.0
cairocffi==0.9.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.4.1
cssselect==1.0.3
cycler==0.10.0
decorator==4.3.0
dill==0.2.8.2
EasyProcess==0.2.5
editdistance==0.5.2
gast==0.2.0
graphviz==0.10.1
grpcio==1.17.1
h5py==2.9.0
html5lib==0.9999999
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.2
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
layer==0.1.14
lxml==4.2.5
Markdown==3.0.1
matplotlib==2.2.3
nltk==3.4
numpy==1.14.5
opencv-python==4.0.0.21
paddlepaddle==1.2.1
pandas==0.23.4
parsel==1.5.1
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prompt-toolkit==2.0.7
protobuf==3.1.0
ptyprocess==0.6.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
PyDispatcher==2.0.5
Pygments==2.3.1
PyHamcrest==1.9.0
pyOpenSSL==18.0.0
pyparsing==2.3.0
pyscreenshot==0.5.1
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
queuelib==1.5.0
rarfile==3.0
recordio==0.1.7
requests==2.9.2
scikit-learn==0.20.1
scipy==1.2.0
Scrapy==1.5.1
service-identity==17.0.0
singledispatch==3.4.0.3
six==1.11.0
sklearn==0.0
tensorboard==1.8.0
tensorflow==1.8.0
termcolor==1.1.0
traitlets==4.3.2
Twisted==18.9.0
urllib3==1.24.1
virtualenv==15.2.0
w3lib==1.19.0
wcwidth==0.1.7
Werkzeug==0.14.1
zope.interface==4.6.0
